+++ esp-hal/src/gpio/lp_io.rs
@@ -197,7 +197,7 @@ macro_rules! lp_gpio {
         $($gpionum:literal)+
     ) => {
         $(
-            impl $crate::gpio::RtcPin for GpioPin<'_, $gpionum> {
+            impl $crate::gpio::RtcPin for paste::paste!($crate::peripherals::[<GPIO $gpionum>]<'_>) {
                 unsafe fn apply_wakeup(&self, wakeup: bool, level: u8) {
                     let lp_io = $crate::peripherals::LP_IO::regs();
                     lp_io.pin($gpionum).modify(|_, w| {
@@ -248,7 +248,7 @@ macro_rules! lp_gpio {
                 }
             }
 
-            impl $crate::gpio::RtcPinWithResistors for GpioPin<'_, $gpionum> {
+            impl $crate::gpio::RtcPinWithResistors for paste::paste!($crate::peripherals::[<GPIO $gpionum>]<'_>) {
                 fn rtcio_pullup(&self, enable: bool) {
                     let lp_io = $crate::peripherals::LP_IO::regs();
                     lp_io.gpio($gpionum).modify(|_, w| w.fun_wpu().bit(enable));